These images depict the front end.
There is a User Information Page.
There is a User login Page (there would be an authentication mechanism using their email)

There is a Game Page that displays the Users participating and their hand (each user can view only their hand). This will display the Turn and the Winner as well. There would be deck at the bottom right and the MatchID at the bottom left,

There is an alternate Game Page for the user where the Users have an avatar (Character) associated with them, a settings and leave button, a A winner indicator, and the card pile. There would also be a played cards pile. 

This is the Minimal Viable Product and is therefore depicted to be quite austere on purpose. If time permits, it shall look much more enchanting, and there may be addional pages (emphasis on MAY).